Added threading using Futures, Changed BigDecimals to Doubles for performance improvement.

+++ src/main/scala/CSPE40824.hw1/Customer.scala
@@ -2,4 +2,4 @@ package CSPE40824.hw1
 
 //TODO add a status attribute?
 /** Att: arrival time, wait time before leaving (deadline), needed service time, id */
-case class Customer(arriveT: BigDecimal, serviceT: BigDecimal, waitT: Int, id: Int)
+case class Customer(arriveT: Double, serviceT: Double, waitT: Double, id: Int)
